app/requirements.txt
fastapi==0.63.0
uvicorn==0.13.3
mysql-connector-python==8.0.23
python-dotenv == 0.17.1
